pyproject.toml
[project]
name = "blueskyapi"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "schroneko", email = "schroneko.fuller@gmail.com" }
]
dependencies = [
    "atproto>=0.0.46",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/blueskyapi"]
